deep learning for computer vision/emotions-detection/requirements.txt
anyio==3.6.2
certifi==2022.9.24
click==8.1.3
coloredlogs==15.0.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
flatbuffers==22.10.26
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mpmath==1.2.1
numpy==1.23.4
onnxruntime==1.13.1
opencv-python-headless
orjson==3.8.1
packaging==21.3
Pillow==9.3.0
protobuf==4.21.9
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
typing_extensions==4.4.0
ujson==5.5.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
